+++ Common/CommonHelper.cs
@@ -226,14 +226,14 @@ namespace Pathoschild.Stardew.Common
         /// <param name="duration">The number of milliseconds during which to keep the message on the screen before it fades (or <c>null</c> for the default time).</param>
         public static void ShowInfoMessage(string message, int? duration = null)
         {
-            Game1.addHUDMessage(new HUDMessage(message, 3) { noIcon = true, timeLeft = duration ?? HUDMessage.defaultTime });
+            Game1.addHUDMessage(new HUDMessage(message, HUDMessage.error_type) { noIcon = true, timeLeft = duration ?? HUDMessage.defaultTime });
         }
 
         /// <summary>Show an error message to the player.</summary>
         /// <param name="message">The message to show.</param>
         public static void ShowErrorMessage(string message)
         {
-            Game1.addHUDMessage(new HUDMessage(message, 3));
+            Game1.addHUDMessage(new HUDMessage(message, HUDMessage.error_type));
         }
 
         /****