+++ NoclipMode/release-notes.md
@@ -2,6 +2,8 @@
 
 # Release notes
 ## Upcoming release
+* Added settings for noclip enabled/disabled notifications.
+* Added optional notification when noclip disabled (off by default).
 * Improved translations. Thanks to LeecanIt (added Italian)!
 
 ## 1.0